Refactor project structure and fix imports
